reduce public footprint a bit more
