+++ src/main/java/lsp/usecase/DistributionTourStartEventHandler.java
@@ -69,7 +69,7 @@ import lsp.LSPCarrierResource;
 		builder.setLogisticsSolutionElement(element);
 		builder.setResourceId(resource.getId());
 		builder.setStartTime(event.getTime());
-		LoggedShipmentTransport transport = builder.build();
+		ShipmentPlanElement transport = builder.build();
 		String idString = transport.getResourceId() + "" + transport.getSolutionElement().getId() + "" + transport.getElementType();
 		Id<ShipmentPlanElement> transportId = Id.create(idString, ShipmentPlanElement.class);
 		lspShipment.getLog().addPlanElement(transportId, transport);
