+++ src/test/java/example/lsp/simulationTrackers/CollectionTrackerTest.java
@@ -2,7 +2,7 @@ package example.lsp.simulationTrackers;
 
 import lsp.*;
 import lsp.controler.LSPModule;
-import lsp.shipment.LSPShipmentImpl;
+import lsp.shipment.ShipmentUtils;
 import org.matsim.contrib.freight.events.eventsCreator.LSPEventCreatorUtils;
 import lsp.LSPInfo;
 import lsp.replanning.LSPReplanningUtils;
@@ -125,7 +125,7 @@ public class CollectionTrackerTest {
 
 		for(int i = 1; i < 2; i++) {
 			Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-			LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+			ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 			Random random = new Random(1);
 			int capacityDemand = random.nextInt(4);
 			builder.setCapacityDemand(capacityDemand);
