+++ src/test/java/lsp/usecase/CollectionLSPSchedulingTest.java
@@ -97,7 +97,7 @@ public class CollectionLSPSchedulingTest {
 
 		for(int i = 1; i < 2; i++) {
         	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
         	Random random = new Random(1);
         	int capacityDemand = random.nextInt(4);
         	builder.setCapacityDemand(capacityDemand);
