+++ src/test/java/lsp/usecase/MultipleShipmentsCompleteLSPSchedulingTest.java
@@ -231,7 +231,7 @@ public class MultipleShipmentsCompleteLSPSchedulingTest {
 		Random rand = new Random(1);
 		 for(int i = 1; i < 100; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = rand.nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
