+++ src/test/java/lsp/usecase/SecondReloadLSPSchedulingTest.java
@@ -193,7 +193,7 @@ public class SecondReloadLSPSchedulingTest {
 		
 		 for(int i = 1; i < 2; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = new Random().nextInt(10);
 	        	builder.setCapacityDemand(capacityDemand);
