+++ src/test/java/lspMobsimTests/CollectionLSPMobsimTest.java
@@ -127,7 +127,7 @@ public class CollectionLSPMobsimTest {
 			ArrayList<Link> linkList = new ArrayList<>( scenario.getNetwork().getLinks().values() );
 			for( int i = 1 ; i < 2 ; i++ ){
 				Id<LSPShipment> id = Id.create( i, LSPShipment.class );
-				LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance( id );
+				ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance( id );
 				//Random random = new Random(1);
 				int capacityDemand = 1 + new Random().nextInt( 4 );
 				builder.setCapacityDemand( capacityDemand );
