+++ src/test/java/lspMobsimTests/CompleteLSPMobsimTest.java
@@ -221,7 +221,7 @@ public class CompleteLSPMobsimTest {
 		Random rand = new Random(1);
 		 for(int i = 1; i < 2; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = 1 +  rand.nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
