+++ src/test/java/lspMobsimTests/FirstReloadLSPMobsimTest.java
@@ -125,7 +125,7 @@ public class FirstReloadLSPMobsimTest {
 		
 		 for(int i = 1; i < 2; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	//Random random = new Random(1);
 	        	int capacityDemand = 1 + new Random().nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
