+++ src/test/java/lspMobsimTests/MultipleIterationsFirstReloadLSPMobsimTest.java
@@ -124,7 +124,7 @@ public class MultipleIterationsFirstReloadLSPMobsimTest {
 		
 		for(int i = 1; i <  1+ numberOfShipments; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	//Random random = new Random(1);
 	        	int capacityDemand = 1 + new Random().nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
