+++ src/test/java/lspMobsimTests/MultipleShipmentsMainRunLSPMobsimTest.java
@@ -164,7 +164,7 @@ public class MultipleShipmentsMainRunLSPMobsimTest {
 		int numberOfShipments = 1 + new Random().nextInt(50);
 		for(int i = 1; i < 1 + numberOfShipments; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = 1 + new Random().nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
