+++ src/test/java/lspMobsimTests/RepeatedMultipleShipmentsCompleteLSPMobsimTest.java
@@ -224,7 +224,7 @@ public class RepeatedMultipleShipmentsCompleteLSPMobsimTest {
 		
 		for(int i = 1; i < 1 +  numberOfShipments; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand =  1 + new Random().nextInt(4);
 	        	builder.setCapacityDemand(capacityDemand);
