+++ src/test/java/lspShipmentAssignmentTests/CollectionLSPShipmentAssigmentTest.java
@@ -7,7 +7,7 @@ import java.util.Collections;
 import java.util.Random;
 
 import lsp.*;
-import lsp.shipment.LSPShipmentImpl;
+import lsp.shipment.ShipmentUtils;
 import lsp.usecase.*;
 import org.junit.Before;
 import org.junit.Test;
@@ -97,7 +97,7 @@ public class CollectionLSPShipmentAssigmentTest {
 
 		for(int i = 1; i < 11; i++) {
         	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
         	int capacityDemand = new Random().nextInt(10);
         	builder.setCapacityDemand(capacityDemand);
