+++ src/test/java/lspShipmentAssignmentTests/CompleteLSPShipmentAssignerTest.java
@@ -5,7 +5,7 @@ import java.util.Collections;
 import java.util.Random;
 
 import lsp.*;
-import lsp.shipment.LSPShipmentImpl;
+import lsp.shipment.ShipmentUtils;
 import lsp.usecase.*;
 import org.junit.Before;
 import org.junit.Test;
@@ -220,7 +220,7 @@ public class CompleteLSPShipmentAssignerTest {
 		
 		 for(int i = 1; i < 11; i++) {
 	        	Id<LSPShipment> id = Id.create(i, LSPShipment.class);
-	        	LSPShipmentImpl.LSPShipmentBuilder builder = LSPShipmentImpl.LSPShipmentBuilder.newInstance(id );
+	        	ShipmentUtils.LSPShipmentBuilder builder = ShipmentUtils.LSPShipmentBuilder.newInstance(id );
 	        	int capacityDemand = new Random().nextInt(10);
 	        	builder.setCapacityDemand(capacityDemand);
